281: move pointsOnSameSide() from globals/utils.js to interpret/utils.js
